[ADD] medismart: added demo data,and  image for doctors and patients
Every time a user starts with a new db, considerable time is spent on creating
data to play around with the module's functionalities. However, the addition of
demo data will prove to be time-efficient.  Images are now added to the
patient's and doctor's form views for obvious reasons. Slight improvement in the
Kanban views of appointments and doctors for better readability.