[ADD] medismart: kanban view for patients and some key  improvements
Added kanban view for patients as it was not there before. Added email and phone
validations in the doctor and patient model using regex which will help restrict
the user from entering random data in these fields. Added SQL constraint in
specialization model to maintain uniqueness. New fields replace the appointment_
date field in the appointment model to better schedule an appointment. Also
added necessary compute and inverse functions along with SQL constraint. Changes
to the demo data ensure that they don't conflict with the newly added constrain-
ts while loading. Added Gantt view for appointments to have an overview of the
overall schedule of doctors. Added photos to the kanban views of doctors and
patients for better identification without opening the record.